+++ nmea_checksum.h
